just realized multiselect plugin is cancer. integrated form saving into text fields. can go ahead with createEventRequest
